+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - ```fix``` Crash in category report when there are transactions without a category.
 - ```fix``` Calculator return 0 when decimals where used in certain languages.
 - ```fix``` Various other crashes.
+- ```fix``` Removed separator between toolbar and categories tabs.
 
 ###Version: 0.11.0
 - ```fix``` Up button is working.